Make Content-Type a symbol to fix POST requests
Any POST request is failing because headers are merged (and cleaned)
using ParamsBuilder. Part of the cleaning process is converting
everything to symbols. So when the request prepares params it checks
the content type but its nil so it defaults to a query string thus
causing POST requests to fail.